Create PredecessorSuccessorinBST
If you want to find successor or predecessor of a particular node in a BST, then simply check if the node.val is >= root.val or not. If it is greater than root.val then it is obvious that the successor won't be in the left subtree of the root as it is a BST. And once you find some node that has value which is greater than p.val then store that node in successor. Also go to the left of that node to confirm that the node you have obtained was the immediate successor of Node p or not?!!
Do the exactly opposite case for predecessor.